Add files via upload
Credit to https://stmorse.github.io/journal/pfr-scrape-python.html for the starter code to create the attributes in order to be searched. This code is able to take any runners yearly stat line and calculate a "rusher rating". This rusher rating is on par to QB rating used by NFL today.